more clean up after review
